extract encoding from file input and move to helper
